+++ src/edcon/profidrive/telegram102.py
@@ -1,14 +1,27 @@
 """Contains code that is related to PROFIDRIVE telegram 102"""
+
 from dataclasses import dataclass, field
 from edcon.profidrive.telegram_base import TelegramBase
-from edcon.profidrive.words import \
-    STW1_SM, NSOLL_B, STW2, MOMRED, G1_STW, ZSW1_SM, \
-    NIST_B, ZSW2, MELDW, G1_ZSW, G1_XIST1, G1_XIST2
+from edcon.profidrive.words import (
+    STW1_SM,
+    NSOLL_B,
+    STW2,
+    MOMRED,
+    G1_STW,
+    ZSW1_SM,
+    NIST_B,
+    ZSW2,
+    MELDW,
+    G1_ZSW,
+    G1_XIST1,
+    G1_XIST2,
+)
 
 
 @dataclass(repr=False)
 class Telegram102(TelegramBase):
     """Holds the implementation of PROFIDRIVE telegram 102"""
+
     # pylint: disable=too-many-instance-attributes
     # Twelve is needed here
     stw1: STW1_SM = field(default_factory=STW1_SM)
@@ -25,17 +38,18 @@ class Telegram102(TelegramBase):
     g1_xist1: G1_XIST1 = field(default_factory=G1_XIST1)
     g1_xist2: G1_XIST2 = field(default_factory=G1_XIST2)
 
-    def input_bytes(self, data: bytes):
-        """Sets the input words from provided byte data"""
-        self.zsw1 = ZSW1_SM.from_bytes(data[0:2])
-        self.nist_b = NIST_B.from_bytes(data[2:6])
-        self.zsw2 = ZSW2.from_bytes(data[6:8])
-        self.meldw = MELDW.from_bytes(data[8:10])
-        self.g1_zsw = G1_ZSW.from_bytes(data[10:12])
-        self.g1_xist1 = G1_XIST1.from_bytes(data[12:16])
-        self.g1_xist2 = G1_XIST2.from_bytes(data[16:20])
+    def inputs(self):
+        """Returns list of input words"""
+        return [
+            self.zsw1,
+            self.nist_b,
+            self.zsw2,
+            self.meldw,
+            self.g1_zsw,
+            self.g1_xist1,
+            self.g1_xist2,
+        ]
 
-    def output_bytes(self) -> bytes:
-        """Returns the byte representation of the output words"""
-        return self.stw1.to_bytes() + self.nsoll_b.to_bytes() + self.stw2.to_bytes() + \
-            self.momred.to_bytes() + self.g1_stw.to_bytes()
+    def outputs(self):
+        """Returns list of output words"""
+        return [self.stw1, self.nsoll_b, self.stw2, self.momred, self.g1_stw]